Дочерний презентер
