+++ src/presenter/main-presenter.js
@@ -1,9 +1,8 @@
-import {render, replace} from '../framework/render.js';
+import {render} from '../framework/render.js';
 import NewPointView from '../view/new-point-view.js';
-import EditPointView from '../view/edit-point-view.js';
-import ListPointView from '../view/list-point-view.js';
 import ContentListView from '../view/content-list-view.js';
 import ListEmptyView from '../view/list-empty-view.js';
+import PointPresenter from './point-presenter.js';
 
 export default class MainPresenter {
   #contentContainer = null;
@@ -32,41 +31,8 @@ export default class MainPresenter {
   };
 
   #renderPoint = (point) => {
-    const pointComponent = new ListPointView(point);
-    const pointEditComponent = new EditPointView(point);
-
-    render(pointComponent, this.#contentList.element);
-
-    const replacePointToForm = () => {
-      replace(pointEditComponent, pointComponent);
-    };
-
-    const replaceFormToPoint = () => {
-      replace(pointComponent, pointEditComponent);
-    };
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        evt.preventDefault();
-        replaceFormToPoint();
-        document.removeEventListener('keydown', onEscKeyDown);
-      }
-    };
-
-    pointComponent.setEditClickHandler(() => {
-      replacePointToForm();
-      document.addEventListener('keydown', onEscKeyDown);
-    });
-
-    pointEditComponent.setFormSubmitHandler(() => {
-      replaceFormToPoint();
-      document.removeEventListener('keydown', onEscKeyDown);
-    });
-
-    pointEditComponent.setEditClickHandler(() => {
-      replaceFormToPoint();
-      document.removeEventListener('keydown', onEscKeyDown);
-    });
+    const pointPresenter = new PointPresenter(this.#contentList.element);
+    pointPresenter.init(point);
   };
 
   #renderEmptyContentList = () => {
